requirements.txt
asgiref==3.5.2
backports.entry-points-selectable==1.1.1
backports.zoneinfo;python_version<"3.9"
certifi==2021.10.8
distlib==0.3.3
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
filelock==3.3.2
importlib-metadata==5.2.0
Markdown==3.4.1
Pillow==9.3.0
pipenv==2021.11.9
platformdirs==2.4.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
virtualenv==20.10.0
virtualenv-clone==0.5.7
zipp==3.11.0
psycopg2-binary==2.9.3
django-cleanup ==6.0.0